+++ examples/run_memory_profiler.sh
@@ -1,0 +1,2 @@
+# Profile script and save to file
+python -m memory_profiler train.py &> memory_profile.txt